requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-authtoken>=1.0,<2.0  # For token-based authentication
django-filter>=2.4,<3.0  # For filtering capabilities (optional)
django-cors-headers>=3.10,<4.0  # For handling CORS (optional)
markdown>=3.3,<4.0  # For Markdown support (optional)
djangorestframework-yasg>=1.20,<2.0  # For Swagger documentation (optional)